Use new homepage gem layout on homepage
As part of the work removing emergency banner from static,
a new gem layout has been created in static for homepage.
This is because the emergency banner has specific styling
when it is present on the homepage.

Previously, the emergency banner homepage styles were applied
by using a parent class selector but this is against the
isolation principle of components. So the emergency banner now
uses a parameter which when set to true applies the correct
styling. This meant that a new gem layout needed to be created
for homepage with this parameter set for the emergency banner
(when the emergency banner is displayed).